requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.7.0
argon2-cffi==23.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
